requirements.txt
# Core ML Libraries
tensorflow>=2.12.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.2.0
pillow>=9.5.0
opencv-python-headless>=4.7.0

# MLflow for experiment tracking
mlflow>=2.4.0

# API Framework
fastapi>=0.100.0
uvicorn>=0.22.0
python-multipart>=0.0.6

# Data versioning
dvc>=3.0.0
dvc[s3]>=3.0.0

# Testing
pytest>=7.3.0
pytest-cov>=4.1.0
httpx>=0.24.0

# Code quality
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0
